Fixed a bug in creating a list of class names.
A list of names for result classes to be loaded was inferred
from all criteria in the 'robo' section under consideration.
Until now this did not pesent a problem since event when in
the study RoboQC section of the configuration file a number
of different criteria sections were present, it was possible
to load all of the AutoQC results for any given entity.

The bug is revealed when lane-level assessment is added to
the RoboQC section under the 'default' configuration. This
type of assessment might require presence of lane-level AutoQC
results for checks that are never performed om sample level,
'tag_metrics' check being the most obvious example.

The fix is to infer the list of class names from the criteria,
which are going to be evaluated for the entity, rather than
from a collection of criteria from the whole RoboQC section.